refactor: Extract xdelta patch functionality into separate module
